+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - updated findProxySourcesFromDDG finish logger info
 - added test for convertBytesToStringSlice
+- added makeDDGSearchURL description
 
 ## 0.3.0 - 2021.08.12